requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nexmo==2.4.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
